Omorganiserer struktur for søkeindeksering for å reflektere at det interne søket nå er legacy
