+++ src/main/resources/lib/search/_legacy/search-repo.ts
@@ -1,23 +1,22 @@
 import * as repoLib from '/lib/xp/repo';
 import { RepoConnection } from '/lib/xp/node';
-import { logger } from '../utils/logging';
-import { runInContext } from '../context/run-in-context';
-import { SEARCH_REPO_ID } from '../constants';
+import { logger } from '../../utils/logging';
+import { runInContext } from '../../context/run-in-context';
+import { SEARCH_REPO_ID } from '../../constants';
 import {
-    getSearchRepoConnection,
-    SEARCH_REPO_CONFIG_NODE,
+    SEARCH_REPO_CONFIG_NODE_LEGACY,
     SEARCH_REPO_CONTENT_BASE_NODE,
     SEARCH_REPO_DELETION_QUEUE_BASE_NODE,
-    SEARCH_REPO_EXTERNAL_CONFIG_NODE,
     SEARCH_REPO_UPDATE_STATE_NODE,
 } from './search-utils';
-import { forceArray } from '../utils/array-utils';
+import { forceArray } from '../../utils/array-utils';
+import { getSearchRepoConnection, SEARCH_REPO_CONFIG_NODE } from '../utils';
 
 const BASE_NODE_KEY = `/${SEARCH_REPO_CONTENT_BASE_NODE}`;
 const DELETION_QUEUE_NODE_KEY = `/${SEARCH_REPO_DELETION_QUEUE_BASE_NODE}`;
 const UPDATE_STATE_NODE_KEY = `/${SEARCH_REPO_UPDATE_STATE_NODE}`;
-const CONFIG_NODE_KEY = `/${SEARCH_REPO_CONFIG_NODE}`;
-const CONFIG_EXTERNAL_NODE_KEY = `/${SEARCH_REPO_EXTERNAL_CONFIG_NODE}`;
+const CONFIG_NODE_KEY = `/${SEARCH_REPO_CONFIG_NODE_LEGACY}`;
+const CONFIG_EXTERNAL_NODE_KEY = `/${SEARCH_REPO_CONFIG_NODE}`;
 
 type UpdateQueueEntry = {
     contentId: string;
@@ -110,12 +109,12 @@ const createBaseNodes = (repo: RepoConnection) => {
         logger.info(`Created node in search repo: ${SEARCH_REPO_UPDATE_STATE_NODE}`);
     }
     if (!repo.exists(CONFIG_NODE_KEY)) {
-        repo.create({ _name: SEARCH_REPO_CONFIG_NODE });
-        logger.info(`Created node in search repo: ${SEARCH_REPO_CONFIG_NODE}`);
+        repo.create({ _name: SEARCH_REPO_CONFIG_NODE_LEGACY });
+        logger.info(`Created node in search repo: ${SEARCH_REPO_CONFIG_NODE_LEGACY}`);
     }
     if (!repo.exists(CONFIG_EXTERNAL_NODE_KEY)) {
-        repo.create({ _name: SEARCH_REPO_EXTERNAL_CONFIG_NODE });
-        logger.info(`Created node in search repo: ${SEARCH_REPO_EXTERNAL_CONFIG_NODE}`);
+        repo.create({ _name: SEARCH_REPO_CONFIG_NODE });
+        logger.info(`Created node in search repo: ${SEARCH_REPO_CONFIG_NODE}`);
     }
 };
