+++ src/main/resources/lib/search/api-handlers/delete-document.ts
