+++ src/main/resources/lib/search/api-handlers/post-document.ts
