+++ src/main/resources/lib/search/document-builder/document-builder.ts
@@ -1,15 +1,15 @@
 import { Content } from '/lib/xp/content';
-import { forceArray } from '../../../utils/array-utils';
-import { getSearchNodeHref } from '../../create-or-update-search-node';
+import { forceArray } from '../../utils/array-utils';
+import { getSearchNodeHref } from '../_legacy/create-or-update-search-node';
 import { generateSearchDocumentId } from '../utils';
 import {
     getContentLocaleRedirectTarget,
     isContentNoIndex,
     isContentPreviewOnly,
-} from '../../../utils/content-utils';
-import { getNestedValues } from '../../../utils/object-utils';
+} from '../../utils/content-utils';
+import { getNestedValues } from '../../utils/object-utils';
 import { getExternalSearchConfig } from '../config';
-import { logger } from '../../../utils/logging';
+import { logger } from '../../utils/logging';
 import {
     SearchDocumentFylke,
     getSearchDocumentFylke,
@@ -18,7 +18,7 @@ import {
 import { SearchDocumentMetatag, getSearchDocumentMetatags } from './field-resolvers/metatags';
 import { getSearchDocumentAudience, SearchDocumentAudience } from './field-resolvers/audience';
 import { getSearchDocumentTextSegments } from './field-resolvers/text';
-import { ContentNode } from '../../../../types/content-types/content-config';
+import { ContentNode } from '../../../types/content-types/content-config';
 import {
     getSearchDocumentContentType,
     SearchDocumentContentType,
@@ -27,7 +27,7 @@ import {
     getSearchDocumentLanguage,
     getSearchDocumentLanguageRefs,
 } from './field-resolvers/language';
-import { isOfficeContent } from '../../../office-pages/types';
+import { isOfficeContent } from '../../office-pages/types';
 import { buildOfficeIngress } from './field-resolvers/office-ingress';
 
 type SearchConfig = Content<'no.nav.navno:search-config-v2'>;
