+++ src/main/resources/lib/search/document-builder/field-resolvers/component-references.ts
@@ -1,10 +1,10 @@
 import { Content } from '/lib/xp/content';
-import { NodeComponent } from '../../../../../types/components/component-node';
-import { ContentNode } from '../../../../../types/content-types/content-config';
-import { getRepoConnection } from '../../../../utils/repo-utils';
-import { getLayersData } from '../../../../localization/layers-data';
-import { CONTENT_ROOT_REPO_ID } from '../../../../constants';
-import { forceArray } from '../../../../utils/array-utils';
+import { NodeComponent } from '../../../../types/components/component-node';
+import { ContentNode } from '../../../../types/content-types/content-config';
+import { getRepoConnection } from '../../../utils/repo-utils';
+import { getLayersData } from '../../../localization/layers-data';
+import { CONTENT_ROOT_REPO_ID } from '../../../constants';
+import { forceArray } from '../../../utils/array-utils';
 
 type ResolvedProductDetails = {
     productDetailsComponents: NodeComponent[];
