+++ src/main/resources/lib/search/event-handlers.ts
@@ -1,12 +1,12 @@
 import * as eventLib from '/lib/xp/event';
 import * as taskLib from '/lib/xp/task';
 import { CONTENT_ROOT_PATH } from '/lib/xp/content';
-import { logger } from '../../utils/logging';
-import { CONTENT_ROOT_REPO_ID, URLS } from '../../constants';
-import { getLayersData } from '../../localization/layers-data';
+import { logger } from '../utils/logging';
+import { CONTENT_ROOT_REPO_ID, URLS } from '../constants';
+import { getLayersData } from '../localization/layers-data';
 import { getExternalSearchConfig, revalidateExternalSearchConfigCache } from './config';
 import { updateExternalSearchDocumentForContent } from './update-one';
-import { isMainDatanode } from '../../cluster-utils/main-datanode';
+import { isMainDatanode } from '../cluster-utils/main-datanode';
 
 let isActive = false;
