--- src/main/resources/lib/search/external/utils.ts
@@ -1,2 +1,0 @@
-export const generateSearchDocumentId = (contentId: string, locale: string) =>
-    `${contentId}-${locale}`;
